Updated version of discord bot

+++ src/main/java/ru/gold_farm_app/income_control/listeners/impl/ServerFractionListenerImpl.java
@@ -15,12 +15,14 @@ public class ServerFractionListenerImpl implements ServerFractionListener {
 
     @Override
     public void onMessageCreate(MessageCreateEvent messageCreateEvent) {
-        if (messageCreateEvent.getMessageContent().startsWith("!server_add")) {
+        if (messageCreateEvent.getMessageContent().startsWith("!server_add") &&
+                messageCreateEvent.getMessageAuthor().getDisplayName().equals("fastrapier1")) {
             String server = messageCreateEvent.getMessageContent().replace(" ", "").substring(11);
             service.addServer(server);
             messageCreateEvent.getChannel().sendMessage("Server " + server + " added!");
         }
-        if (messageCreateEvent.getMessageContent().startsWith("!server_delete")) {
+        if (messageCreateEvent.getMessageContent().startsWith("!server_delete") &&
+                messageCreateEvent.getMessageAuthor().getDisplayName().equals("fastrapier1")) {
             String server = messageCreateEvent.getMessageContent().replace(" ", "").substring(14);
             service.deleteServer(server);
             messageCreateEvent.getChannel().sendMessage("Server " + server + " deleted!");
